Added account modifier and delete account functionalities to mobile app
Added a check controller for user and refresh tokens.
Changed stops ids in local DB according with remote one.
Fixed an issue in StatsActivity.
Fixed an error in WifiUtility.
Refactorized Utils.

+++ BusTrack/BusTrack/Data/Line.cs
@@ -17,6 +17,6 @@ namespace BusTrack.Data
         public RealmList<Stop> stops { get; }
         [JsonProperty("stops")]
         [Ignored]
-        public List<int> stopIds { get; set; }
+        public List<long> stopIds { get; set; }
     }
 }
